[Refactor] bubblemenu에서 아이템 분리
